Fix api schema mixup in revocation routes
