+++ aries_cloudagent/revocation/routes.py
@@ -270,7 +270,7 @@ class RevokeRequestSchema(CredRevRecordQueryStringSchema):
     )
 
 
-class PublishRevocationsSchemaAnoncreds(OpenAPISchema):
+class PublishRevocationsSchema(OpenAPISchema):
     """Request and result schema for revocation publication API call."""
 
     rrid2crid = fields.Dict(
@@ -293,7 +293,7 @@ class TxnOrPublishRevocationsResultSchema(OpenAPISchema):
     """Result schema for credential definition send request."""
 
     sent = fields.Nested(
-        PublishRevocationsSchemaAnoncreds(),
+        PublishRevocationsSchema(),
         required=False,
         metadata={"definition": "Content sent"},
     )
@@ -330,7 +330,7 @@ class ClearPendingRevocationsRequestSchema(OpenAPISchema):
     )
 
 
-class CredRevRecordResultSchemaAnoncreds(OpenAPISchema):
+class CredRevRecordResultSchema(OpenAPISchema):
     """Result schema for credential revocation record request."""
 
     result = fields.Nested(IssuerCredRevRecordSchema())
@@ -613,7 +613,7 @@ async def revoke(request: web.BaseRequest):
 
 
 @docs(tags=["revocation"], summary="Publish pending revocations to ledger")
-@request_schema(PublishRevocationsSchemaAnoncreds())
+@request_schema(PublishRevocationsSchema())
 @querystring_schema(CreateRevRegTxnForEndorserOptionSchema())
 @querystring_schema(RevRegConnIdMatchInfoSchema())
 @response_schema(TxnOrPublishRevocationsResultSchema(), 200, description="")
@@ -686,7 +686,7 @@ async def publish_revocations(request: web.BaseRequest):
 
 @docs(tags=["revocation"], summary="Clear pending revocations")
 @request_schema(ClearPendingRevocationsRequestSchema())
-@response_schema(PublishRevocationsSchemaAnoncreds(), 200, description="")
+@response_schema(PublishRevocationsSchema(), 200, description="")
 async def clear_pending_revocations(request: web.BaseRequest):
     """Request handler for clearing pending revocations.
 
@@ -1070,7 +1070,7 @@ async def update_rev_reg_revoked_state(request: web.BaseRequest):
     summary="Get credential revocation status",
 )
 @querystring_schema(CredRevRecordQueryStringSchema())
-@response_schema(CredRevRecordResultSchemaAnoncreds(), 200, description="")
+@response_schema(CredRevRecordResultSchema(), 200, description="")
 async def get_cred_rev_record(request: web.BaseRequest):
     """Request handler to get credential revocation record.
